Fix design flaw that made accessing private channels impossible
Turns out JDA and the Discord API don't allow you to get a list of your open
conversations. In fact, you can only ever have 10 open conversations so
this means using the snowflake of a private channel is useless since we have
no way of locating the channel and may even need to reopen it in which case
we would need to know the user.

Therefore, these changes introduce a prefix system for all channels used by
Discord. When resolving channel identifier, the prefix is used to identify
private message channels and reopen them, if required.