Remove flag for verbose IRC logging, IRC will now always log verbose via new logger
With the capability to configure the logging at runtime, it becomes quite handy to be able to enable verbose
logging of IRC traffic. Therefore the logger code was moved into a new class so we can specifically configure
it's logger.

This allows us to debug protocol level problems with IRC more easily