Clarify error message in escape network command
If the user attempts to add a new graph with an already existing name, the parser will interpret that action as an attempt to modify an existing edge and thous take a parsing path that will ultimate consume the entire graph specification. Since this string is then fed into the edge specification parser, it will fail to parse the entire graph specification as an edge specification, resulting in an error message, telling the user that the given edge specification is malformed. While is this technically the correct error message, if the user expected a new graph to be generated they will assume that there is an issue with the graph definition, not realizing that they attempted to call the modify branch. This commit adds an additional check that will warn the user if a modification command contains the separation marker and could thous be a graph definition.