Handle more broad errors in Timer Command
There appears to be a rare case in which the TimerCommand will stop fetching new times, indicating an issue with the periodic retrival of new timers. In an attempt to debug this issue, we are going to broaden the scope of the enclosing catch block to hopefully catch this bug.